:sparkles: attr code base

+++ src/main/java/com/gxk/jvm/instruction/IfIcmpGtInst.java
@@ -3,9 +3,9 @@ package com.gxk.jvm.instruction;
 import com.gxk.jvm.rtda.Frame;
 
 public class IfIcmpGtInst implements Instruction {
-  public final short offset;
+  public final int offset;
 
-  public IfIcmpGtInst(short offset) {
+  public IfIcmpGtInst(int offset) {
     this.offset = offset;
   }
 
@@ -19,7 +19,7 @@ public class IfIcmpGtInst implements Instruction {
     Integer val2= frame.operandStack.popInt();
     Integer val1= frame.operandStack.popInt();
     if (val1 > val2) {
-      frame.nextPc = (int) offset;
+      frame.nextPc = offset;
     }
   }
 }
